+++ tests/Parser/Visitor/Example/DocBlock/AnonymousClassDocBlockArgument.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace J6s\PhpArch\Tests\Parser\Visitor\Example\DocBlock;
+
+class AnonymousClassDocBlockArgument
+{
+}